+++ src/lib.rs
@@ -3,11 +3,12 @@ Prometheus instrumentation for Rocket applications.
 
 # Usage
 
-Add this crate to your `Cargo.toml`:
+Add this crate to your `Cargo.toml` alongside Rocket 0.5.0-rc.1:
 
 ```toml
 [dependencies]
-rocket_prometheus = "0.9.0"
+rocket = "0.5.0-rc.1"
+rocket_prometheus = "0.10.0-rc.1"
 ```
 
 Then attach and mount a [`PrometheusMetrics`] instance to your Rocket app: