+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.10.0-rc.1] - 2022-02-10
+### Changed
+
+- Upgrade Rocket dependency to 0.5.0-rc.1, and make the fairing async.
 - Remove an allocation on every request when getting the `status` label, provided the status code is between 100 and 999.
 
 ## [0.9.0] - 2021-10-17
@@ -74,7 +78,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First version of the crate released to crates.io.
 
-[Unreleased]: https://github.com/sd2k/rocket_prometheus/compare/v0.9.0...HEAD
+[Unreleased]: https://github.com/sd2k/rocket_prometheus/compare/v0.10.0-rc.1...HEAD
+[0.10.0-rc.1]: https://github.com/sd2k/rocket_prometheus/compare/v0.9.0...v0.10.0-rc.1
 [0.9.0]: https://github.com/sd2k/rocket_prometheus/compare/v0.8.1...v0.9.0
 [0.8.1]: https://github.com/sd2k/rocket_prometheus/compare/v0.8.0...v0.8.1
 [0.8.0]: https://github.com/sd2k/rocket_prometheus/compare/v0.7.0...v0.8.0